% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-add_dy_dt.R
\name{add_dy_dt}
\alias{add_dy_dt}
\title{Creates a Correlation Matrix}
\usage{
add_dy_dt(uids)
}
\arguments{
\item{df}{dataframe in format}

\item{db}{dataframe in format}
}
\value{
Returns a correlation matrix
}
\description{
This takes a raw data file and creates a systems correlation database
}
\examples{

}
\author{
David Hammond
}
