% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-add_meta.R
\name{add_meta}
\alias{add_meta}
\title{Creates a Correlation Matrix}
\usage{
add_meta(df, folder = "data")
}
\arguments{
\item{df}{dataframe in format}

\item{folder}{dataframe in format}
}
\value{
Returns a correlation matrix
}
\description{
This takes a raw data file and creates a systems correlation database
}
\examples{

}
\author{
David Hammond
}
