% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-centrality_db.R
\name{centrality_db}
\alias{centrality_db}
\title{Creates a Correlation Matrix}
\usage{
centrality_db(rval = 0.5)
}
\arguments{
\item{df}{dataframe in format}

\item{use_geocode_and_time_as_obs}{dataframe in format}
}
\value{
Returns a correlation matrix
}
\description{
This takes a raw data file and creates a systems correlation database
}
\examples{

}
\author{
David Hammond
}
