% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentage_change_db.R
\name{change_db}
\alias{change_db}
\title{Creates a Correlation Matrix}
\usage{
change_db(df, time_col, group_by_col, value_col)
}
\arguments{
\item{df}{dataframe in format}
}
\value{
Returns a correlation matrix
}
\description{
This takes a raw data file and creates a systems correlation database
}
\examples{

}
\author{
David Hammond
}
