% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-changes_db.R
\name{changes_db}
\alias{changes_db}
\title{Creates a Correlation Matrix}
\usage{
changes_db(df, folder)
}
\arguments{
\item{df}{dataframe in format}

\item{newscale}{dataframe in format}
}
\value{
Returns a correlation matrix
}
\description{
This takes a raw data file and creates a systems correlation database
}
\examples{

}
\author{
David Hammond
}
