% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-corr_cluster_db.R
\name{corr_cluster}
\alias{corr_cluster}
\title{Creates a Correlation Matrix}
\usage{
corr_cluster(folder, num_clust = 100)
}
\arguments{
\item{df}{dataframe in format}

\item{use_geocode_and_time_as_obs}{dataframe in format}
}
\value{
Returns a correlation matrix
}
\description{
This takes a raw data file and creates a systems correlation database
}
\examples{

}
\author{
David Hammond
}
