% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlate_db.R
\name{corr_db}
\alias{corr_db}
\title{Creates a Correlation Matrix}
\usage{
corr_db(df, use_geocode_and_time_as_obs = T)
}
\arguments{
\item{df}{dataframe in format}

\item{use_geocode_and_time_as_obs}{dataframe in format}
}
\value{
Returns a correlation matrix
}
\description{
This takes a raw data file and creates a systems correlation database
}
\examples{

}
\author{
David Hammond
}
