% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-get_cluster.R
\name{get_cluster}
\alias{get_cluster}
\title{Creates a Correlation Matrix}
\usage{
get_cluster(x, whch = 1)
}
\arguments{
\item{df}{dataframe in format}

\item{use_geocode_and_time_as_obs}{dataframe in format}
}
\value{
Returns a correlation matrix
}
\description{
This takes a raw data file and creates a systems correlation database
}
\examples{

}
\author{
David Hammond
}
