% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-get_functions.R
\name{get_corpus}
\alias{get_corpus}
\title{Creates a Correlation Matrix}
\usage{
get_corpus()
}
\arguments{
\item{df}{dataframe in format}

\item{folder}{dataframe in format}
}
\value{
Returns a correlation matrix
}
\description{
This takes a raw data file and creates a systems correlation database
}
\examples{

}
