% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-file_structure.R
\name{get_db}
\alias{get_db}
\title{Creates a Correlation Matrix}
\usage{
get_db(folder, db)
}
\arguments{
\item{db}{dataframe in format}

\item{df}{dataframe in format}
}
\value{
Returns a correlation matrix
}
\description{
This takes a raw data file and creates a systems correlation database
}
\examples{

}
\author{
David Hammond
}
