% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-granger_db.R
\name{granger_db}
\alias{granger_db}
\title{Creates a Correlation Matrix}
\usage{
granger_db(test = F, subset_granger = c("goal.x", 16))
}
\arguments{
\item{df}{dataframe in format}

\item{use_geocode_and_time_as_obs}{dataframe in format}
}
\value{
Returns a correlation matrix
}
\description{
This takes a raw data file and creates a systems correlation database
}
\examples{

}
\author{
David Hammond
}
