% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-granger_execute.R
\name{granger_execute}
\alias{granger_execute}
\title{Creates a Correlation Matrix}
\usage{
granger_execute(gcode, systr_file)
}
\arguments{
\item{df}{dataframe in format}

\item{use_geocode_and_time_as_obs}{dataframe in format}
}
\value{
Returns a correlation matrix
}
\description{
This takes a raw data file and creates a systems correlation database
}
\examples{

}
\author{
David Hammond
}
