% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-setup.R
\name{syssetup}
\alias{syssetup}
\title{Creates a Correlation Matrix}
\usage{
syssetup(df, newscale = c(1, 10))
}
\arguments{
\item{df}{dataframe in format}

\item{newscale}{dataframe in format}
}
\value{
Returns a correlation matrix
}
\description{
This takes a raw data file and creates a systems correlation database
}
\examples{

}
\author{
David Hammond
}
