% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-granger_db.R
\name{systr_granger}
\alias{systr_granger}
\title{Creates a Correlation Matrix}
\usage{
systr_granger(corpus)
}
\arguments{
\item{df}{dataframe in format}

\item{use_geocode_and_time_as_obs}{dataframe in format}
}
\value{
Returns a correlation matrix
}
\description{
This takes a raw data file and creates a systems correlation database
}
\examples{

}
\author{
David Hammond
}
