% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-plot_granger_timeseries.R
\name{systr_granger_time_series}
\alias{systr_granger_time_series}
\title{Creates a Correlation Matrix}
\usage{
systr_granger_time_series(x, which_row = 1)
}
\arguments{
\item{df}{dataframe in format}

\item{use_geocode_and_time_as_obs}{dataframe in format}
}
\value{
Returns a correlation matrix
}
\description{
This takes a raw data file and creates a systems correlation database
}
\examples{

}
\author{
David Hammond
}
