% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12-indicator_summary.R
\name{systr_indicator_summary}
\alias{systr_indicator_summary}
\title{Creates a Correlation Matrix}
\usage{
systr_indicator_summary(indicator, rval = 0.5, pval = 0.1)
}
\arguments{
\item{df}{dataframe in format}

\item{use_geocode_and_time_as_obs}{dataframe in format}
}
\value{
Returns a correlation matrix
}
\description{
This takes a raw data file and creates a systems correlation database
}
\examples{

}
\author{
David Hammond
}
