% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-setup.R
\name{systr_setup}
\alias{systr_setup}
\title{Creates a Correlation Matrix}
\usage{
systr_setup(df, meta, folder = "data", newscale = c(1, 5))
}
\arguments{
\item{df}{dataframe in format}

\item{folder}{dataframe in format}

\item{newscale}{dataframe in format}
}
\value{
Returns a correlation matrix
}
\description{
This takes a raw data file and creates a systems correlation database
}
\examples{

}
\author{
David Hammond
}
