% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-trends_db.R
\name{trends_together_db}
\alias{trends_together_db}
\title{Creates a Correlation Matrix}
\usage{
trends_together_db(folder)
}
\arguments{
\item{folder}{dataframe in format}
}
\value{
Returns a correlation matrix
}
\description{
This takes a raw data file and creates a systems correlation database
}
\examples{

}
\author{
David Hammond
}
