% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-what_correlates.R
\name{what_correlates}
\alias{what_correlates}
\title{Creates a Correlation Matrix}
\usage{
what_correlates(uids, folder, rval = 0.5)
}
\arguments{
\item{uids}{dataframe in format}

\item{rval}{dataframe in format}
}
\value{
Returns a correlation matrix
}
\description{
This takes a raw data file and creates a systems correlation database
}
\author{
David Hammond
}
