% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-whats_the_difference_between.R
\name{whats_the_difference_between}
\alias{whats_the_difference_between}
\title{Creates a Correlation Matrix}
\usage{
whats_the_difference_between(cluster1, cluster2, signif = 0.05)
}
\arguments{
\item{cluster1}{dataframe in format}

\item{cluster2}{dataframe in format}

\item{signif}{dataframe in format}

\item{df}{dataframe in format}
}
\value{
Returns a correlation matrix
}
\description{
This takes a raw data file and creates a systems correlation database
}
\author{
David Hammond
}
