% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-plot_granger_graph.R
\name{systr_granger_graph}
\alias{systr_granger_graph}
\title{Creates a Correlation Matrix}
\usage{
systr_granger_graph(
  granger,
  pval = 0.15,
  filter_for_same_direction = T,
  plot = T
)
}
\arguments{
\item{df}{dataframe in format}

\item{use_geocode_and_time_as_obs}{dataframe in format}
}
\value{
Returns a correlation matrix
}
\description{
This takes a raw data file and creates a systems correlation database
}
\examples{

}
\author{
David Hammond
}
